ADD zoom effect on camera UPDATE SceneSpawner script so it's more modular UPDATE health potion so it syncs with network UPDATE EnemyFollow now enables/disables the NavMeshAgent if it's Mine on network ADD health pickup sound

+++ El Llamo 3D/Assets/Scripts/Player/Health.cs
@@ -189,7 +189,7 @@ public class Health : MonoBehaviour
 
     private void CreateRagdoll()
     {
-        NetworkClient.Instance.LastSpawner.SpawnForNonPlayer(NonPlayerIndexes.Ragdoll_Corpse, transform.position, Quaternion.identity);
+        NetworkClient.Instance.LastSpawner.SpawnForNonPlayer((int)NonPlayerIndexes.Ragdoll_Corpse, transform.position, Quaternion.identity);
         /*
         SWNetworkMessage msg = new SWNetworkMessage();
         msg.Push(transform.position);
